magit--minibuf-default-add-commit: Fix position of commit
Functions for reading commits let-bind minibuffer-default-add-function
to magit--minibuf-default-add-commit in order to add the commit at
point to the minibuffer's "next history".  However, the added commit
isn't actually accessible via M-n, at least with default completion.
Instead an 'M-n M-n M-p' sequence is needed to reach it.

The problem is that goto-history-element expects the value returned by
minibuffer-default-add-function to include minibuffer-default, and, if
there is a non-nil default that's not the same as the commit,
magit--minibuf-default-add-commit puts the commit upstream of the next
position that goto-history-element selects.

Fix this issue by making minibuffer-default-add-function insert
minibuffer-default in the front of the list, in the same way as
minibuffer-default-add-completions does.  When there is a default
other than the commit at point, the commit is accessible via 'M-n
M-n'.

Notes on two third-party completion frameworks:

 * With ivy, the let-bound minibuffer-default-add-function doesn't
   seem to have an effect, while using minibuffer-with-setup-hook, as
   is done in the Emacs code base, does.  That's worth considering as
   a future change.

 * With helm, the minibuffer-selected-window call in
   magit--minibuf-default-add-commit selects a helm window rather than
   the expected window, so the commit isn't returned.  We could work
   around that by binding the magit-commit-at-point return value
   outside of the lambda, though that might be a bit too much special
   handling for one completion framework.

Fixes #4740.