git-rebase-mode-show-keybindings: fix key display for mapped commands
This is a regression from 5ddf251 which addressed a font lock change in
Emacs 28.1. That commit routed `git-rebase-command-descriptions' through
`substitute-command-keys' to propertize them but commands (and therefore
the keys) in `git-rebase-*' were then rendered, for example, as \'c'.

This patch routes all the mapped commands through
`substitute-command-keys' so that users have a consistent rendering:
either all of the keys are not font locked (Emacs 27) or they are font
locked (Emacs 28 and beyond).