magit-save-repository-buffers: Handle let-bound default-directory
When `default-directory' is let-bound around a call to git, then that
binding is still in effect when `save-some-repositories' (and thus
this function) process the current buffer.  (For other buffers that is
not the case because `default-directory' is a buffer-local variable.)

As a result, if the current buffer is modified and visits a file in
another repository, then this function wrongly concluded that the file
is located in the repository specified by `default-directory'.

If the wip modes are enabled, then saving a tracked file causes wip
refs to be updated.  When that happens, the let-binding is still in
effect, so an attempt is made to stage the visited file in the wrong
repository, which results in an error being raised, about the file
being outside the repository.

Counter this by internally let-binding `default-directory' based on
`buffer-file-name'.