diff: Remove partially broken and unnecessary kludge
A comment claimed that "git diff" ignores "--no-prefix" for new and
renames files, but that is no longer true.  It possibly was never
true.  I tried confirm that it isn't true for the oldest git release
that we still support, v2.2.0, but I am no longer able to build that.

I have decided to remove the kludge for "git diff" because the kludge
would not have worked for new files anyway.  There were no bug reports
concerning that, so it is very likely that this kludge really isn't
needed.

The kludge is still needed for "git log -L", but has to be improved.