magit-commit-squash-internal: Restore window conf after showing diff
When magit-commit-show-diff is non-nil, the diff window is typically
deleted by with-editor-return restoring the initial window
configuration, but it isn't deleted by commands that go through
magit-commit-squash-internal.  In the case of
magit-commit{,-instant}-{fixup,squash}, this is because the editor
isn't invoked.  For magit-commit-augment, the editor is invoked, but
the diff is already displayed by the time with-editor stores the
initial window configuration.

Teach magit-commit-squash-internal to restore the window configuration
so that these commands behave the same as other commit commands.

Closes #4767.