magit-display-buffer: inhibit-same-window for magit-display-buffer-noselect
I prefer to use display-buffer-same-window by default via
display-buffer-alist. It is the most predictable buffer switching
behavior.

But there're some cases, e.g. magit-commit, magit-blame that needs to
show the diff in side window. In this case, we pass `t` as
`display-buffer` ACTION argument to inhibit displaying in the same
window.