test: Set protocol.file.allow=always to restore old Git behavior
Under Git v2.38.1, the magit-get and magit-toplevel:submodule tests
fail due to clone errors.  Both of these tests add a submodule with a
_local_ URL.  As of a1d4f67c12 (transport: make `protocol.file.allow`
be "user" by default, 2022-10-01) [*], Git prevents submodule
initialization from "file://" (and bare local) sources by default for
security reasons.

In the context of Magit's test suite, this isn't a security risk, so
teach magit-with-test-directory to add the file protocol to the list
of allowed protocols.

[*] This commit was brought into Git's maint-* branches for v2.30.X
    through 2.38.X.