Consider Git's exec path when searching for extensions
Commands that depend on optional third-party Git extensions use
executable-find to check whether the extension is available.  That
misses the legitimate case where an extension is installed in Git's
exec path rather than a directory in $PATH.

Update these spots to call a new helper that looks in Git's exec path
before calling executable-find.  Note that the order is chosen to
follow Git's precedence (though that doesn't matter for any of the
current callers because they don't use the return value).

Closes #4812.