git-commit-post-finish-hook-timeout: New option
`git-commit-post-finish-hook' is not run until the new commit has been
created. If Git takes more than 1 second to do so, the hook isn’t run.

Whereas a 1 second timeout seems reasonable for most use cases, there
are use cases where a greater time is desirable.

For ex., when git is configured to sign commits with GnuPG, the user
might be interactively prompted to enter the passphrase of the GnuPG
private key.
If the user takes more than 1 second to type their passphrase, the
hook isn’t run.

In order to address such use cases, the new variable
`git-commit-post-finish-hook-timeout' allows to tune the value of this
timeout.

Co-authored-by: Jonas Bernoulli <jonas@bernoul.li>
  Name of new option and docstring tweaks.