magit-{branch-or-,}commit-at-point: Consider removal blame chunks
magit-commit-at-point and magit-branch-or-commit-at-point extract the
commit from a blame chunk with

  (magit-current-blame-chunk 'addition t)

For a magit-blame-reverse chunk, 'git blame' is called again to get an
'addition' chunk, leading to the returned commit confusingly not
matching the chunk's commit.  A downstream consequence of this is that
magit-show-commit displays the addition commit for a removal chunk
(while magit-diff-show-or-scroll-up shows the expected removal commit,
because it doesn't rely on magit-branch-or-commit-at-point).

As of 4b8eab3a (magit-{branch-or,}commit-at-point: Only use blame
chunk when blaming, 2022-05-27), magit-commit-at-point and
magit-branch-or-commit-at-point invoke magit-current-blame-chunk only
when Blame mode is enabled, so it's safe to assume that there is an
existing blame chunk.  Rely on it to get the appropriate commit.

Fixes #4834.