magit-log-merged: Fix incorrect calculation of distance from tip
When calculating the distance of a commit that is actually on a branch
from the tip of that branch (i.e., the <N> such that branch~<N>
identifies that commit), we should only count the path along first
parents, as that is how the syntax branch~<N> is defined. Before this
commit, we were counting _all_ commits between the commit and the tip
of the branch, which is more than the correct number, in the case that
the history between commit and branch is non-linear. Thus, the point
around which we log was chosen incorrectly, and an incorrect set of
commits was shown in the resulting log.