magit-pull: support `--no-rebase` argument to `git pull`.
In recent versions of git, non-rebase merges are no longer the
default. Instead, when trying to pull divergent branches, if
`--no-rebase` is not passed explicitly (and if the default has not
been overridden in config), an error message is printed to the effect
that the user must specify how to deal with the divergent branch.

This change allows the user to explicitly specify that the divergent
branches should be merged.