Inhibit use of libgit by default
The support for libgit in magit is only a stub.  At this time, the
only reason valid reason to use it, is when one works on extending
support.  There is no harm in using it even when not working to
improve it, but there is a risk of breakage if we don't explicitly
disable it by default.

More precisely, if the emacs library `libgit' is available but the
`libegit2' module has not been compiled, then the user is asked
whether they want to compile it.  If that fails, e.g., because no
compiler is available, then they will be asked every time they
restart Emacs.

Closes #4880.