magit-auto-revert-mode: Leave disabled in one more case
If, during initialization, `magit-autorevert' loaded without
`global-auto-revert-mode' already being enabled, then the default
value of `magit-auto-revert-mode' becomes t.  If later, but still
during initialization, `global-auto-revert-mode' is enabled, then
that default for `magit-auto-revert-mode' is not appropriate and
we have to explicitly disable it.

Usually it should not be necessary to loag Magit during initialization
and `global-auto-revert-mode' is something one would expect to be
enabled early on, if so desired; but still, this is something that
can happen and we should try to handle it as users would expect.

Note that we didn't already intend to do that before this commit.
We however used to do something similar for the obsolete and since
removed `magit-revert-buffers' variable.  In [1: 2253379bf7] we
removed that variable.

1: 2018-06-05 2253379bf72e97371774add0aae71cc4b57a69a1
   magit-revert-buffers: Remove deprecated variable

Closes #4891.