magit-log-set-commit-limit: Modify copy of magit-buffer-log-args
magit-log-set-commit-limit calls `delete' on the magit-buffer-log-args
value.  This may be the same object as the magit-log-default-arguments
property.  For example, with the default
magit-direct-use-buffer-arguments, magit-log-buffer-file passes the
result of (get 'magit-log-mode 'magit-log-default-arguments) to
magit-log-setup-buffer.  In that case, calling
magit-log-set-commit-limit in the resulting log deletes the -nN
argument from the default log arguments.

Use `remove' instead to avoid modifying magit-log-default-arguments.

Fixes #4898.