magit-xref-setup: Limit number of items
We already tried to do so but since this call to
`called-interactively-p' always returned nil, that did not actually
happen.  We tried to only do it "when called interactively", because
that is what `help-setup-xref' does.  I see no reason to make this
distinction here, so trim the history unconditionally.  However
limiting to just ten items seems too extreme, so limit to 30.