Complete switch to new magit-completing-read-multiple
In [1: c10e10c9a4] we added `magit-completing-read-multiple*' as a
replacement for `magit-completing-read-multiple'.  The goal was to
eventually remove the latter and reuse its name for the former.

Do that now.  Well, one kludge remains: for now the function, whose
name ends with "*", is preserved as an alias for the function, whose
name does not.

1: 2021-06-29 c10e10c9a427c192dec841e524a6d3d47d2b21ee
   Update our completing-read-multiple replacement/wrapper