magit-insert-remote-branches: Fix Git 2.40 compatibility
If magit-insert-remote-branches sees a remote HEAD line in the
for-each-ref output, it notes what value the symref points to so that
it can mark that branch with magit-branch-remote-head.  Before storing
the value, it asserts that the refname:short value for the line
matches the expected value, "<remote>/HEAD".

This assertion fails as of Git's 613bef56b8
(shorten_unambiguous_ref(): avoid sscanf(), 2023-02-15), part of the
v2.40.0 release.  "refs/remotes/<remote>/HEAD" is now shortened to
"<remote>" rather than "<remote>/HEAD", matching the fallback behavior
described in 'man gitrevisions'.

Switch the assertion over to using the full reference to make it
compatible with Git versions before and after v2.40.0.

Closes #4903.