Improve docstring of magit alias
The purpose of this alias is to make `magit-status' -- the main
entry point -- discoverable for complete beginners who have read
zero documentation.  See #3290.

Optimize that line of the docstring for that purpose.  Depending
on the used completion framework and customization, that might be
shown in the minibuffer after, e.g., "M-x m a g TAB".

I have also contemplated removing this alias because it gets in the
way of having, e.g., "M-x m a g TAB" complete to "magit-" instead of
merely "magit".  I might still do that, but for now I have decided
that it is not enough of an issue when using a suitable completion
framework and settings.