Remove useless magit-branch-description-check-buffers
The value of functions on `find-file-hook' has no relevance and
this function has no side-effects.  It should have been removed
in [1: c07b475051], instead of merely being renamed and losing
the side-effect it used to have.  It used to refresh the popup
buffer, which became irrelevant when we switched to `transient'
in that commit.

1: 2019-02-14 c07b47505115564d8c79b988fe204844d8da3b3f
   Define magit-branch-configure as a transient command