Deprecate magit-git-dir to avoid needlessly repeating work
Five years ago in [1: 5f407a29ac] we took the first step toward this.

1: 2018-10-15 5f407a29ac63b338c3c75e30b4633b44d5745036
   magit-gitdir: New function

Previously we cached the wrong thing.  `expand-file-name' is cheap but
"git rev-parse --git-dir" is not.  Because `magit-git-dir' is where we
cached, rev-parse was still needlessly called multiple times with the
same arguments during a refresh.  The effect should be most noticeable
during a rebase.