Support displaying errors that occur when washing a section
`magit-git-wash' silently removes a section when the command produces
no output on stdout and stderr is ignored.  For a section among many,
that is usually the right thing to do, but when the section is the
only or primary section of a buffer, then we might want to display
the error (see subsequent commits).

Add a new function `magit--git-wash', which takes a new argument that
allows doing that.  Also add variable `magit--git-wash-keep-error' to
allow forcing that for all sections inserted with `magit-git-wash'.
That is experimental and might be removed again.