Show errors inline in log buffers
Only do it for log sections that make up the whole buffer.
That makes it necessary to add a new argument, and because
I would like for it to go first, that is a breaking change.

Take this opportunity to replace `magit-insert-log' with
`magit--insert-log' for symmetry with `magit--insert-diff',
which is is a function on `magit-diff-sections-hook', run by
`magit-diff-refresh-buffer'.  `magit-log-refresh-buffer' runs
no hook.  Instead it calls `magit--insert-log' directly.