Ignore magit-buffer-topdir when preparing another buffer
Since [1: 623d335b4d] we set `magit-buffer-topdir' and use that in
`magit-toplevel', but when we determine and create a buffer that
concerns another buffer, then that value does not apply.

Without this, visiting a module from the status buffer of the
super-repository, resulted in the super-repositories buffer being
reused to display the module's status, which then caused many
commands to fail when invoked in that directory.

1: 2023-03-17 623d335b4dddc64f205748862520ed41661a7a80
   Memorize gitdir and topdir for each Magit buffer