Only insert stderr into log and diff buffers on failure
In [1: 78a979fde5] we added support for inserting git's stderr inline
when washing a section and in [2: 2bd3db69d6] and [3: d05b1ed381] we
started using that.  However, that did not take into account that git
may output warnings while still succeeding.  In that case we must not
insert the warnings because the section washers are not prepared to
deal with them.

1: 2023-03-18 78a979fde52815242b165f083d171259db28e0b4
   Support displaying errors that occur when washing a section

2: 2023-03-18 2bd3db69d60e90cf4f4139a2bdae5bb448685e37
   Show errors inline in diff buffers

3: 2023-03-18 d05b1ed381c44d53dfe16a964169ecbdae9067ff
   Show errors inline in log buffers