magit-wip-mode: Speedup activation by avoiding duplicated work
Reduce work to one call to `magit-toplevel' for each directory
from which one or more files are being visited, and to one call
to `magit-tracked-files' for each repository from which one or
more files are being visited.

Unfortunately we cannot do the same if `magit-wip-after-save-mode'
is enabled directly, but that is only provided as a separate mode
for backward compatibility anyway.

Closes #4768.