magit-wash-bisect-log: Ignore '# status:' lines
As of Git v2.37.0, bisect outputs a status line if either the good or
bad revision isn't specified.  For example:

  $ git bisect start
  status: waiting for both good and bad commits

If a user starts a bisect from the command line and then visits the
repo with Magit, these status lines trigger a parse error.  (Starting
a bisect from Magit never emits these status messages because both
revisions are always specified upfront.)

There's not much value in parsing and displaying this info, so teach
magit-wash-bisect-log to detect and delete these lines when processing
the 'git bisect log' output.

Fixes #4787.