Introduce customization to `magit-insert-worktrees'
A common workflow when using worktrees is to have one worktree per
feature-/fix-branch. In this workflow, it can be more useful to
provide details on the ticket being worked rather than the location of
the worktree (which might even be generated on the ticket ID).

Introduce `magit-worktree-section-details-function' to allow
customizing the details inserted into worktree sections -- extracting
the default behavior into `magit-insert-worktree-path'.