Introduce and use more robust worktree-list wrappers
The function for this purpose, `magit-list-worktrees', is limited to
providing a fairly sparse set of information about each worktree: the
worktree path, its current HEAD and checked-out branch, and whether or
not it's bare. This isn't sufficient for other tooling (or future
Magit functionality) that wants to use other properties of the
worktree (e.g., whether or not it's prunable).

Introduce `magit-list-worktrees-raw' to generically capture every
property reported by Git as a string-valued hashmap. The
post-processing that had been present in `magit-list-worktrees' is
moved to `magit-list-worktrees-fix-worktree' so that it can be mapped
over `magit-list-worktrees-raw'.

`magit-list-worktrees' is re-implemented in terms of these two
functions to keep the existing interface for legacy code.

Care is taken in the implementation of `magit-list-worktrees-raw' to
avoid unnecessary string allocations by using `string-match-p' and
`substring' instead of `string-split'.