magit-restore-window-configuration: Select buffer
Apparently calling `set-window-configuration' does not update the
current buffer, so we have to do that ourselves, using code that
looks highly suspicious.

If we don't do that, then `current-buffer' returns the wrong buffer in
`magit-section-post-command-hook'.  This is not fatal for most code in
that function, and thus went unnoticed, but it trips up the recently
added `cursor-sensor-move-to-tangible'.

Closes #4940.