Add user option 'git-commit-usage-hint'
The intent is to make it easier for users to visualise the common key
bindings of git-commit.el.  New users can benefit from the option to
place the hint in the header-line, as they tend to miss what is shown
in the echo area.

The idea for this change was discussed here:
<https://github.com/magit/magit/discussions/4942>.