git-commit-setup: Ensure misleading message is shadowed
`git-commit-usage-message' is long enough to
potentially span more than one line.  Previously one had to remove
`with-editor-usage-message' from `git-commit-setup-hook' to prevent
that.  But then `with-editor-usage-message' wasn't displayed either
and that meant that the user got to see the misleading message that
`server-execute' insists on showing.

Add a new mechanism for disabling showing the long message; setting
`git-commit-usage-message' to nil.