Place comma after "e.g." and "i.e."
We already do this in Transient's documentation because that's what
is being done in Emacs' documentation and Transient is part of Emacs.
Let's be consistent even at the cost of making the English unhappy.

Also add a comma before "e.g." and "i.e.", if not inside parentheses,
which is the correct thing to do, regardless which style guide one
chooses to follow.