magit-stash-{apply,pop}: Fix case when there are staged changes
It is not possible to preserve a stash index when applying it and
there are staged changes.  Previously we just always tried to preserve
it anyway, using the `--index' argument, and only failed we fell back
to applying without that argument.

That used to work, but due to a change/regression in Git that stopped
working quite some time ago.  The futile attempt to use `--index'
causes the staged changes to be unstaged and when there are unstaged
changes (to files also touched by the stash), then it is not possible
to apply a stash at all.

(Users still have to be aware that they might have to stage before
being able to apply a stash, despite the presence of uncommitted
changes.)