magit-stash-pop: Use pop instead of apply
Restores pop behavior lost in
bdade489ead0ffd1de0709042d5166837efea4e8.
Previously, "git stash apply" was used when popping, with an explit
drop due to handling of the "--index" argument. Now that "--index" is
passed conditionally, and the drop was removed, the call here can be
replaced by the expected "git-stash-pop".