Use when-let instead of --when-let
In a few cases use `when-let*' or `and-let*' instead.

I haven't done a full 180 on anaphoric conditionals; if a builtin
equivalent existed, I would use that in some instances.

However, having looked at all the existing uses of `when-let', it
turns out there are way fewer such cases than I had anticipated.