Don't preserve visibility in log buffers
Looking up the visibility of thousands of sections is very slow.
In log buffers the visibility usually also does not matter, because
commits usually do not have any children.  When showing diffs or
other additional data, then visibility does matter, and refreshing
would now reset the visibility to the initial state, but usually
one does not have to refresh a log buffer, when hiding and showing
details about some commits.