Don't set comment-use-syntax to nil.
Back when commit messages were always entered in `text-mode`, this was
reasonable.  Now that `git-commit-major-mode` allows us to type our
comments in different modes, we need to respect the comment style of
whichever mode we are in.  This affects how `auto-fill-mode` can wrap
lines.

The default for `comment-use-syntax` is to detect if the current major
mode supports comments.  If so, it will use that.  If not, it will use
the comment character that is specified in git's config (usually `#`).

This fixes #4990