git-commit-setup-font-lock: Set comment-end to the empty string
Some major-modes used while editing commit messages, may set that
variable and other `comment-*' variables to values that are not
compatible with Git's understanding of comments, which always only
uses a comment character at the beginning, either the default "#" or
what is specified using `core.commentchar'.  Git's understanding of
how comments are delimited takes precedence over the major-modes
style, because it is Git that strips the comments.

`markdown-mode', for example, sets `comment-start' to "<!--" and
`comment-end' to "-->"; and we failed to override the latter until
now.

Furthermore, `markdown-fill-paragraph' does not handle Git-style
comments, which is to be expected, but forces us neutralize it when
called on a comment.  This is not necessary for any of the other
modes we suggest for `git-commit-major-mode'.

Closes #4990.
Replaces #4991.