magit-revert-rev-file-buffer: Run major-mode hooks again
This reverts [1: f331092df4], which disabled it for the benefit of
`lsp', which does something that isn't compatible with what we are
doing here.

At least for the time being, advice the `lsp' function instead,
which I have been told, also works around the incompatibility.

Fixes #4986.
Re #4683.
Re doomemacs/doomemacs#6309.

1: 2022-05-08 f331092df4d4dfc0a2a7424d929a9c845088d57f
   magit-revert-rev-file-buffer: Use delay-mode-hook