Remove text properties when copying from magit buffers
If there is yanked text with the `magit-section' property at point,
many magit commands would believe that there indeed is a section at
point.  This may or may not result in an error, and if not, then the
resulting behavior can be very confusing.

It might have been nice to preserve the `face' property, and maybe
some others, but `remove-list-of-text-properties' does not support
prefixing LIST-OF-PROPERTIES with `not' to remove all but the named
properties.