Ensure a seq version with seq-keep is used
For older Emacs releases we depend on an updated `seq' release from
GNU ELPA, for `seq-keep'.  Unfortunately something else may already
have required `seq', before `package' had a chance to put the more
recent version earlier on the `load-path', and thus the outdated
built-in version might be loaded.  If that is the case, unload it
and require `seq' again, to load the new version.

If you are wondering whether it is worth the trouble, keep in mind
that going forward we might want to use other additions in the future.
Also I really want to use this function.  I was the one who suggested
its addition because it was one of the few significant gaps in `seq'
and `cl-lib' in comparison to `dash'.

Closes #5011.