magit-section-{pre,post}-command-hook: Don't suppress in transient
In [1: 95b432530a], we started suppressing these functions when a
transient is active.  The main concern of that commit are context-
menus, and its message makes no mention of why we did this for
transients.

I have now run into cases, where suppressing this is undesirable
(i.e., when a transient prefix binds movement suffixes or allows
non-suffixes).

1: 2022-03-25 95b432530a6b131e63bec7411ee768f1dee9f629
   Do not set point when invoking context-menu