magit-hunk-section-map: Honor smerge-command-prefix again
We stopped doing that in [1: deb10e984e] without providing an
explanation.

I am undecided whether honoring/hi-jacking this variable is
appropriate.  Its docstring says that it concerns `smerge-mode'
commands, but our `magit-smerge-*' commands are used in `magit-*-mode'
buffers.  It is likely that a user would want to change the prefix in
both keymaps; but if they don't, then the undocumented side-effect is
confusing.

Additionally I have reservations about such variables in the first
place.  They make certain very specific, customization slightly
easier, at the cost of making widespread binding changes harder.

I am going back to using this variable in this context, to revert
a change in behavior, but if we had never done this before, then
I would most likely reject a proposal to start doing so now.

Closes #5021.

1: 2023-02-13 deb10e984e16201182b0569f7df7d30ec3b8afa9
   Use defvar-keymap for keymaps containing context menus