git-commit: Ignore leading comment when inserting headers
It is suggested in several places on the internet to use a git
commit template, for instance here [1].  These often include a
header comment:

    # Title: Summary, imperative, start upper case, don't end with a period
    # No more than 50 chars. #### 50 chars is here:  #

[1]: https://gist.github.com/lisawolderiksen/a7b99d94c92c6671181611be1641c733

This change ensures that headers are inserted after this leading
comment (meaning there are no empty lines before it), instead of
above it.  Headers are still inserted before the comments added
by Git itself.