magit-insert-am-sequence: Account for already applied patches
magit-insert-am-sequence inserts a section for all patches in
$GIT_DIR/rebase-apply/ and then displays ORIG_HEAD through HEAD via
magit-sequence-insert-sequence.  If the 'git am' call fails to apply a
patch, this leads to repeated sections for successfully applied
patches because Git leaves the patch files around after successfully
applying a patch.

Avoid the repeated sections by explicitly inserting a section only for
$GIT_DIR/rebase-apply/ patches that have not been applied, leaving the
rest to the magit-sequence-insert-sequence call.

Closes #5024.