magit-refresh-buffer: Define as a command
For some `magit-mode' buffer `magit-refresh' is not suitable, and
in those cases, `magit-refresh-buffer' is a good substitute.

Some buffers whose mode derives from `magit-mode' (for example
`forge-topic-mode' buffers) are *not* necessarily tied to a Git
repository, so `magit-refresh' cannot be used, since that expects
to be run inside a Git repository.

Additionally, while setting `magit-refresh-status-buffer' to nil
would also keep `magit-refresh' from refreshing the status buffer,
making `magit-refresh-buffer' available as a command is a more
flexible mechanism.