Favor transient-prefix-object over transient--prefix
This function just returns the value of
  (or transient--prefix transient-current-prefix)

and was added so that authors of transient commands would not have
to worry about when to use which variable.  Since many will refer
to Magit as a source of examples, I do here what I recommend to keep
things simple.

I'm leaving the uses of `transient--prefix' in "magit-transient.el"
intact, because there, they are used in methods, and implementing
your own methods is considered advanced usage.