magit-cherry-refresh-buffer: Don't inhibit use of markers
In [1: 241b327b7a] we started to forgo using markers in log buffers
for performance reasons.  In other log buffers that works well because
they contain nothing but the log and after that has been washed, the
buffer is not modified.  Cherry buffers are modified later on, to
insert the count of cherry commits, which invalidates all higher
positions.

In the cherry buffer, go back to using markers, which get
automatically adjusted.

We could instead drop the cherry count, but that would be a more
noticeable change.  If there are so many cherries that the use of
markers becomes a performance issue, then I expect that effect to be
dwarfed by the cost of determining the cherries in the first place.

1: 2023-08-05 241b327b7ae04e898cdc6343fbee0a1be3b2ce3f
   Don't use markers in log buffers