Avoid unnecessary commit graph calculations
From my debugging, `range' here is usually something like
"HEAD~10..HEAD" (up to the value of `magit-log-section-commit-count'),
which of course means 'find me everything reachable from HEAD that
isn't reachable from HEAD~10'. We also pass "-n10" as a further
limiter. This appears to be duplicative, so prefer "-n10" alone.

"HEAD~10..HEAD" not only requires Git to parse HEAD~10, but also
requires reachability logic at each step when walking the graph.
Performance of this reachability logic is improved with tools like
`git-commit-graph', but it is not what I'd consider 'solved'.

On a branch with ~3400000 commits (~3700000 commits in --all), this is
a significant improvement (tested with `magit-refresh-verbose'):

current implementation:

    magit-insert-recent-commits    0.884695 !!
    magit-insert-recent-commits    0.508423 !!
    magit-insert-recent-commits    0.357324 !!
    magit-insert-recent-commits    0.360272 !!
    magit-insert-recent-commits    0.356748 !!

new implementation:

    magit-insert-recent-commits    0.125902 !!
    magit-insert-recent-commits    0.125966 !!
    magit-insert-recent-commits    0.125386 !!
    magit-insert-recent-commits    0.125143 !!
    magit-insert-recent-commits    0.125195 !!

This change has also been tested on a repository with <10 commits and
had no issues that I could see.