git-commit-turn-on-flyspell: Don't check below cut-line
When `commit.verbose' is set to true in git's configuration, a special
line will be added to the (temporary) commit message.  All content below
this will be ignored by git when constructing the eventual commit
message.  This is defined as `cut_line' in wt-status.c in git's source:

https://github.com/git/git/blob/3e0d3cd5c7def4808247caf168e17f2bbf47892b/wt-status.c#L38-L39

I noticed that, for exceptionally large commits, magit was triggering
a commensurately large flyspell operation.  This appears to be a small
oversight in the logic of `git-commit-turn-on-flyspell'.

This commit attempts fix the performance issue by instructing flyspell
to ignore all lines after `cut_line'.