magit-branch: Add global variable suffixes by default
It was strange that one could already set `branch.<branch>.pushRemote'
directly from this menu, but to set `remote.pushDefault', one had to
use the "configure a particular branch" menu.  One even had to choose
which branch to configure, in order to configure this repository-wide
setting.

In `magit-branch' we cannot use "b" for `magit-update-default-branch',
because that conflicts with the binding for `magit-checkout'.  So use
"B" instead, both here and in `magit-branch-configure' for consistency.