magit-insert-section--create: Find keymap by name here
This also has to fix an inconsistency regarding the value of the
`keymap' slot, which is now always a symbol.  The actual keymap is
now consistently determined at the last possible moment, right before
setting the `keymap' text property.  Previously this was only done at
that time, if the value originated from the class's `:initform', but
not if it was looked up by name.