Make discouraged use of git-commit-mode-hook safer and more convenient
`git-commit-mode' is a minor-mode, as opposed to being the major-mode
used in commit message buffers (as its name may suggest, and many users
seem to assume).

We therefore decided a long time ago that `git-commit-hook-mode' should
not be used and provided `git-commit-setup-hook' as an alternative.

Despite that, users keep trying to use it anyway, and often that leads
to issues.  It doesn't help that this hook has a generic generated
docstring, which does nothing to discourage its use.

Instead of continuing to tell users not to use this hook, one by one, or
taking more drastic measures to prevent its use, just make it an alias
for `git-commit-setup-hook'.  (It is also too late to instead remove
*that* hook.)

Prevent `git-commit-mode-hook' from being run when `git-commit-mode' is
enabled.  It is now run a bit later under the `git-commit-setup-hook'
name, while also demoting errors.  Demoting errors is important, because
errors that occur here, are (1) hard to debug (because there is no
backtrace), (2) common (due to user errors), and (3) they outright
prevent committing.

It also makes a difference that the hook is run slightly later now.
Some things that would fail at the earlier time, actually work at the
later time.  For example, functions that rely on font-lock being setup
(which is reasonable to do if you wrongly assume that `git-commit-mode'
is a major-mode), as many seem to do, only works at the later time.

Also move the call to `git-commit-mode' itself further down.

In summary, it seems safest *and* most convenient, to just let users
add to the same hook under two different names, and to run it at the
appropriate time, while demoting errors.