magit-{browse,visit}-thing: Handle raw URL links
Every once in a while, someone suggests that it should be possible to
follow URL links (often those displayed by user hooks and thus in the
process buffer), which I used to decline, on the grounds that it would
be way too costly to do that, for too little gain, and that the
workaround (using `browse-url') was simple and obvious enough.

By adding support for urls to `magit-browse-thing' (because that is
where it belongs, according to Magit's browse|visit distinction) and
`magit-visit-thing' (because that (i.e., "RET") is what users are
likely to try using), we get the desired behavior -- without having
to spend time looking for urls ahead while washing process output.

So now "RET" visits links, unless `magit-visit-thing' is remapped to
something else, that is.  The mouse still cannot be used.

Closes #5134.